// client/src/interfaces/dashboardInterfaces/dashboardInterfaces.tsx
export interface DigitalDisplay {
  digitalDisplayId: string; // Unique identifier for the digital display. TODO: Will be generated by the backend when a new display is created.
  title: string; // The title of the digital display, which can be set by the user. It will be displayed on the card header.
  varName: string; // represents the name of the variable that the display is linked to. This should correspond to a variable in the data structure table.
  varValue: string; // This will hold the current value of the variable linked to the display.
  radioId: number | null; //Represents the id of the radio that the varName is linked to.
  type: string; // Represents the type of the varName
  posx: number; //Represents the x position of the display on the dashboard grid.
  posy: number; //Represents the y position of the display on the dashboard grid.
}

export interface DigitalDisplayProps {
  displays: DigitalDisplay[];
  setDisplays: React.Dispatch<React.SetStateAction<DigitalDisplay[]>>;
  radios: any[]; // TODO:This should be typed according to the structure of a radio in your application.
}